refactor(fintoc): Restrutcture entire folder structure to have better navigability and more intuitive naming
