+++ lib/fintoc/movements/client/client.rb
@@ -1,0 +1,10 @@
+require 'fintoc/base_client'
+require 'fintoc/movements/client/links_methods'
+
+module Fintoc
+  module Movements
+    class Client < BaseClient
+      include LinksMethods
+    end
+  end
+end
