+++ lib/fintoc/movements/resources/transfer_account.rb
@@ -1,4 +1,4 @@
-require 'fintoc/resources/movements/institution'
+require 'fintoc/movements/resources/institution'
 
 module Fintoc
   module Movements
