+++ lib/fintoc/transfers/client/client.rb
@@ -1,0 +1,10 @@
+require 'fintoc/base_client'
+require 'fintoc/transfers/client/entities_methods'
+
+module Fintoc
+  module Transfers
+    class Client < BaseClient
+      include EntitiesMethods
+    end
+  end
+end
