+++ lib/fintoc/transfers/resources/entity.rb
