+++ spec/lib/fintoc/transfers/client_spec.rb
@@ -1,0 +1,28 @@
+require 'fintoc/transfers/client/client'
+
+RSpec.describe Fintoc::Transfers::Client do
+  let(:api_key) { 'sk_test_9c8d8CeyBTx1VcJzuDgpm4H-bywJCeSx' }
+  let(:client) { described_class.new(api_key) }
+
+  describe '.new' do
+    it 'creates an instance of TransfersClient' do
+      expect(client).to be_an_instance_of(described_class)
+    end
+  end
+
+  describe '#to_s' do
+    it 'returns a formatted string representation' do
+      expect(client.to_s)
+        .to include('Fintoc::Transfers::Client')
+        .and include('🔑=')
+    end
+  end
+
+  it 'responds to transfers-specific methods' do
+    expect(client)
+      .to respond_to(:get_entity)
+      .and respond_to(:get_entities)
+  end
+
+  it_behaves_like 'a client with entities methods'
+end
