+++ spec/spec_helper.rb
@@ -2,6 +2,11 @@ require 'bundler/setup'
 require 'fintoc'
 require 'vcr'
 
+# Load shared examples
+Dir[File.join(File.dirname(__FILE__), 'support', 'shared_examples', '**', '*.rb')].each do |f|
+  require f
+end
+
 VCR.configure do |c|
   vcr_mode = /rec/i.match?(ENV.fetch('VCR_MODE', nil)) ? :all : :once
