+++ spec/vcr/Fintoc_Movements_Client/behaves_like_a_client_with_links_methods/links_functionality/_get_account/get_a_linked_account.yml
