+++ spec/vcr/Fintoc_Movements_Client/behaves_like_a_client_with_links_methods/links_functionality/_get_accounts/prints_accounts_to_console.yml
