+++ spec/vcr/Fintoc_Movements_Client/behaves_like_a_client_with_links_methods/links_functionality/_links/get_all_the_links_from_a_given_link_token.yml
